requirements.txt
torch
open_clip_torch
flask
pillow
numpy
faiss-cpu
faiss-gpu